项目管理系统/auto_test/requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-html>=4.1.0
pyyaml>=6.0.1
openpyxl>=3.1.2
urllib3>=2.0.0
